Gemma/requirements.txt
numpy
sounddevice
edge-tts
playsound==1.2.2
faster-whisper
llama-cpp-python
easyocr
opencv-python
torch
torchvision
argostranslate
pygame